# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ServiPro API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: users
  - name: providers
  - name: bookings
  - name: ratings
  - name: reports
  - name: stats
  - name: notifications

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"

  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      security:
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          description: Radius in kilometers
          schema:
            type: number
        - name: category
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderProfile"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /ratings:
    post:
      operationId: createRating
      tags: [ratings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"

  /ratings/user/{id}:
    get:
      operationId: getUserRatings
      tags: [ratings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingSummary"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /reports/{id}/status:
    patch:
      operationId: updateReportStatus
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportStatusInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [client, provider]
        category:
          type: string
        region:
          type: string
      required: [name, email, password]

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        category:
          type: string
        region:
          type: string
      required: [id, name, email, role, status]

    UpdateUserStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, pending, suspended]
      required: [status]

    ProviderProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        category:
          type: string
        region:
          type: string
        isAvailable:
          type: boolean
        averageRating:
          type: number
        distanceKm:
          type: number
      required: [id, name, email]

    CreateBookingInput:
      type: object
      properties:
        providerId:
          type: integer
        serviceType:
          type: string
        scheduledAt:
          type: string
        lat:
          type: number
        lng:
          type: number
        region:
          type: string
        notes:
          type: string
      required: [providerId, serviceType]

    UpdateBookingStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, in_progress, completed, cancelled]
      required: [status]

    Booking:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        providerId:
          type: integer
        serviceType:
          type: string
        status:
          type: string
        scheduledAt:
          type: string
        region:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required: [id, clientId, providerId, serviceType, status]

    CreateRatingInput:
      type: object
      properties:
        bookingId:
          type: integer
        ratedId:
          type: integer
        score:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [bookingId, ratedId, score]

    Rating:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        raterId:
          type: integer
        ratedId:
          type: integer
        score:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required: [id, bookingId, raterId, ratedId, score]

    RatingSummary:
      type: object
      properties:
        userId:
          type: integer
        averageScore:
          type: number
        totalRatings:
          type: integer
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/Rating"
      required: [userId, averageScore, totalRatings, ratings]

    CreateReportInput:
      type: object
      properties:
        reportedId:
          type: integer
        bookingId:
          type: integer
        reason:
          type: string
      required: [reportedId, reason]

    UpdateReportStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [open, under_review, resolved]
        adminNotes:
          type: string
      required: [status]

    Report:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: integer
        reportedId:
          type: integer
        bookingId:
          type: integer
        reason:
          type: string
        status:
          type: string
        adminNotes:
          type: string
        createdAt:
          type: string
      required: [id, reporterId, reportedId, reason, status]

    PlatformStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalProviders:
          type: integer
        bookingsToday:
          type: integer
        pendingReports:
          type: integer
        totalBookings:
          type: integer
        completedBookings:
          type: integer
      required: [totalUsers, totalProviders, bookingsToday, pendingReports, totalBookings, completedBookings]

    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, title, body]