# lib/api-zod/src/generated/types/createRatingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ServiPro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRatingInput {
  bookingId: number;
  ratedId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  comment?: string;
}